requirements.txt
Pillow
numpy
streamlit_drawable_canvas
pytest-shutil
requires.io
pybase64